requirements.txt
python-dotenv==1.0.1
python_fasthtml==0.4.5
pytz==2022.1
supabase==2.7.3